LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.moveTo.nonfinite.worker-expected.txt


PASS moveTo() with Infinity/NaN is ignored
